+++ packages/react/DataTable/src/components/DataTableBulkActions/dataTableBulkActions.types.ts
@@ -1,0 +1,15 @@
+import { ReactNode, HTMLAttributes } from "react";
+
+import { CheckboxProps } from "@nimbus-ds/components";
+
+export interface DataTableBulkActionsProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
+  /** Properties of the checkbox element rendered in the Bulk Actions component */
+  checkbox: Omit<CheckboxProps, "label" | "id">;
+  /** Optional link element rendered next to the Bulk Actions controller */
+  link?: ReactNode;
+  /** Action component that controls the Bulk Actions */
+  action: ReactNode;
+  /** Lable for the checkbox element */
+  label: string;
+}