+++ packages/react/DataTable/src/components/DataTableFooter/index.ts
@@ -1,0 +1,4 @@
+import { DataTableFooter } from "./DataTableFooter";
+
+export { DataTableFooter } from "./DataTableFooter";
+export default DataTableFooter;