+++ packages/react/DataTable/src/components/DataTableHeader/index.ts
@@ -1,0 +1,4 @@
+import { DataTableHeader } from "./DataTableHeader";
+
+export { DataTableHeader } from "./DataTableHeader";
+export default DataTableHeader;