+++ packages/react/DataTable/src/index.ts
@@ -1,0 +1,5 @@
+import { DataTable } from "./DataTable";
+
+export { DataTable } from "./DataTable";
+export type { DataTableProps } from "./dataTable.types";
+export default DataTable;