+++ packages/react/ProductUpdates/src/productUpdates.spec.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { render, screen } from "@testing-library/react";
+import { render, screen, fireEvent, act } from "@testing-library/react";
 
 import { ProductUpdates } from "./ProductUpdates";
 import { ProductUpdatesProps } from "./productUpdates.types";
@@ -31,31 +31,46 @@ const makeSut = (
 
 describe("GIVEN <ProductUpdates />", () => {
   describe("WHEN rendered", () => {
-    it("SHOULD render title text correctly", () => {
+    it("SHOULD render title text correctly", async () => {
       makeSut({});
+      await act(() => {
+        fireEvent.mouseEnter(screen.getByTestId("product-updates-element"));
+      });
       expect(screen.getByText(titleText)).toBeDefined();
     });
 
-    it("SHOULD render content text correctly", () => {
+    it("SHOULD render content text correctly", async () => {
       makeSut({});
+      await act(() => {
+        fireEvent.mouseEnter(screen.getByTestId("product-updates-element"));
+      });
       expect(screen.getByText(contentText)).toBeDefined();
     });
 
     describe("SHOULD render the closing elements correctly", () => {
-      it("SHOULD render the default X button", () => {
+      it("SHOULD render the default X button", async () => {
         makeSut({});
+        await act(() => {
+          fireEvent.mouseEnter(screen.getByTestId("product-updates-element"));
+        });
         expect(screen.getByTestId("icon-element")).toBeDefined();
       });
 
-      it("SHOULD render the dismiss link correctly", () => {
+      it("SHOULD render the dismiss link correctly", async () => {
         makeSut({ dismissLink: "Dismiss link" });
+        await act(() => {
+          fireEvent.mouseEnter(screen.getByTestId("product-updates-element"));
+        });
         expect(screen.getByText("Dismiss link")).toBeDefined();
       });
     });
 
-    it("SHOULD render body content correctly", () => {
+    it("SHOULD render body content correctly", async () => {
       makeSut({
         bodyContent: <p data-testid="body-content-element">Body content</p>
+      });
+      await act(() => {
+        fireEvent.mouseEnter(screen.getByTestId("product-updates-element"));
       });
       expect(screen.getByTestId("body-content-element")).toBeDefined();
     });