+++ packages/react/CalloutCard/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This component is used to provide the user with useful information, tips, tutorials or shortcuts to other parts of the product. It is a visual component specially designed to be used in contexts such as integration, tutorials or dashboards, where we need to show different relevant information in the same space.
 
+## 2023-03-02 `1.4.1`
+
+### 📚 3rd party library updates
+
+- Updated `@nimbus-ds/components@2.10.0`. ([#22](https://github.com/TiendaNube/nimbus-patterns/pull/22) by [@juanchigallego](https://github.com/juanchigallego))
+- Updated `@nimbus-ds/styles@6.13.0`. ([#22](https://github.com/TiendaNube/nimbus-patterns/pull/22) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-02-17 `1.4.0`
 
 ### 🎉 New features